requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
colorama==0.4.6
httpx[socks] 
aiohttp==3.8.1 